# sql/pgwire: avoid new memory allocation in writeTextDatum(Date)