@nvanbenschoten 2019-10-07
The overwhelming majority of the remaining time is spent in `fmt.Fprintf`:

![Screen Shot 2019-10-07 at 4 10 00 PM](https://user-images.githubusercontent.com/5438456/66344952-faf18880-e91c-11e9-8b7d-3a7cfed4e38a.png)

Unless we want to get very manual and do something like `time.Time.AppendFormat`, which contains custom logic in [`appendInt`](https://github.com/golang/go/blob/a0894ea5b5c326f1ddc7c4c6674d5858f8761dc8/src/time/format.go#L363), I don't think there's much more to do here. We'll need to just accept the remaining perf regression.